most updated code, and it works
It seems to cover the basic functionality of avoiding obstacles exactly how we wanted it to, but it still doesn't control the main fan, and we don't have the drive fan yet so that obviously isn't implemented yet either. Testing will definitely need to be done to determine the precise value of some numbers, like how far from a wall it should start turning.